Exclude settings to a header file without impl
